Complement the README.md

+++ debian-based/ansible/cleaner/clean-docker.sh
@@ -1,5 +1,23 @@
 #!/bin/bash
 
 sudo apt -y purge docker-ce docker-ce-cli containerd.io
-sudo rm -vrf /var/lib/docker
-sudo rm -vrf /var/lib/containerd
+
+if [ -d /var/lib/docker ]; then
+    sudo rm -vrf /var/lib/docker
+fi
+
+if [ -d /var/lib/containerd ]; then
+    sudo rm -vrf /var/lib/containerd
+fi
+
+if [ -f /etc/apt/sources.list.d/docker.list ]; then
+    sudo rm -v /etc/apt/sources.list.d/docker.list
+fi
+
+if [ -f /usr/share/keyrings/docker-archive-keyring.gpg ]; then
+    sudo rm -v /usr/share/keyrings/docker-archive-keyring.gpg
+fi
+
+if [ -f /etc/docker/daemon.json ]; then
+    sudo rm -v /etc/docker/daemon.json
+fi
